Changed the "Long" value to "Double" in the Result class. Made the necessary changes after that added Nodes for "+","-","*","/","max","min","%" operations added Casting overload for "+" added Negation overload for "-"

+++ json-logic-java/src/main/java/com/northernwall/jsonLogic/JsonLogicTree.java
@@ -75,7 +75,7 @@ public class JsonLogicTree {
         JsonToken token = jsonReader.peek();
         switch (token) {
             case NUMBER:
-                temp.put(name, new Result(jsonReader.nextLong()));
+                temp.put(name, new Result(jsonReader.nextDouble()));
                 return;
             case BOOLEAN:
                 temp.put(name, new Result(jsonReader.nextBoolean()));
